refactor: move saved schedule logic to custom hook
Moves state and logic from SavedSchedulesPage to useSavedSchedules
hook for better organization and reusability.
