+++ app/saved/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useState, useEffect, useRef } from 'react';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { AlertCircle, X, Download } from 'lucide-react';
@@ -15,114 +14,38 @@ import {
   AlertDialogTitle,
 } from '@/components/ui/alert-dialog';
 import { WeeklySchedule } from '@/components/schedule/weekly-schedule';
-import { useScheduleManagement } from '@/lib/hooks/use-schedule-management';
 import { ImportScheduleDialog } from '@/components/saved/import-schedule-dialog';
 import { ShareDialog } from '@/components/saved/share-dialog';
 import { EmptyState } from '@/components/saved/empty-state';
 import { ScheduleGrid } from '@/components/saved/schedule-grid';
 import { SelectedCourseList } from '@/components/saved/selected-course-list';
 import { useSavedSchedules } from '@/lib/hooks/use-saved-schedules';
-import type { SavedSchedule } from '@/lib/services/schedules';
-import type { Course } from '@/lib/types/course';
 
 export default function SavedSchedulesPage() {
-  const previewRef = useRef<HTMLDivElement>(null);
-  // --- Hooks ---
   const {
+    previewRef,
     schedules,
     isLoading,
     error,
-    deleteSchedule,
-    importSchedule,
-    updateSharing,
+    selectedCourses,
+    conflicts,
+    showDeleteAlert,
+    setShowDeleteAlert,
+    scheduleToDelete,
+    activeSchedule,
+    showImportDialog,
+    setShowImportDialog,
+    showShareDialog,
+    setShowShareDialog,
+    shareUrl,
+    handleDeleteClick,
+    handleConfirmDelete,
+    handlePreviewClick,
+    closePreview,
+    handleExport,
+    handleShareClick,
+    handleImport,
   } = useSavedSchedules();
-
-  const { selectedCourses, conflicts, setSelectedCoursesDirectly } =
-    useScheduleManagement();
-
-  // --- UI State ---
-  const [showDeleteAlert, setShowDeleteAlert] = useState(false);
-  const [scheduleToDelete, setScheduleToDelete] =
-    useState<SavedSchedule | null>(null);
-  const [activeSchedule, setActiveSchedule] = useState<SavedSchedule | null>(
-    null
-  );
-  const [showImportDialog, setShowImportDialog] = useState(false);
-  const [showShareDialog, setShowShareDialog] = useState(false);
-  const [shareUrl, setShareUrl] = useState('');
-
-  // --- Effects ---
-  useEffect(() => {
-    if (activeSchedule) {
-      setSelectedCoursesDirectly(activeSchedule.schedule_data || []);
-      setTimeout(() => {
-        previewRef.current?.scrollIntoView({
-          behavior: 'smooth',
-          block: 'start',
-        });
-      }, 100);
-    } else {
-      setSelectedCoursesDirectly([]);
-    }
-  }, [activeSchedule, setSelectedCoursesDirectly]);
-
-  // --- Handlers ---
-  const handleDeleteClick = (schedule: SavedSchedule) => {
-    setScheduleToDelete(schedule);
-    setShowDeleteAlert(true);
-  };
-
-  const handleConfirmDelete = async () => {
-    if (!scheduleToDelete) return;
-    const success = await deleteSchedule(scheduleToDelete.id);
-    if (success && activeSchedule?.id === scheduleToDelete.id) {
-      setActiveSchedule(null);
-    }
-    setShowDeleteAlert(false);
-    setScheduleToDelete(null);
-  };
-
-  const handlePreviewClick = (schedule: SavedSchedule) => {
-    setActiveSchedule(schedule);
-  };
-
-  const closePreview = () => {
-    setActiveSchedule(null);
-  };
-
-  const handleExport = (schedule: SavedSchedule) => {
-    const exportData = {
-      type: 'planify-schedule',
-      version: 1,
-      scheduleName: schedule.schedule_name,
-      data: schedule.schedule_data,
-    };
-    const dataStr = JSON.stringify(exportData, null, 2);
-    const dataUri =
-      'data:application/json;charset=utf-8,' + encodeURIComponent(dataStr);
-    const exportFileDefaultName = `${schedule.schedule_name}.json`;
-    const linkElement = document.createElement('a');
-    linkElement.setAttribute('href', dataUri);
-    linkElement.setAttribute('download', exportFileDefaultName);
-    linkElement.click();
-  };
-
-  const handleShareClick = async (schedule: SavedSchedule) => {
-    let targetSchedule = schedule;
-    if (!targetSchedule.is_shared) {
-      const updatedSchedule = await updateSharing(targetSchedule.id, true);
-      if (!updatedSchedule) return; // Error handled in hook
-      targetSchedule = updatedSchedule;
-    }
-    const url = `${window.location.origin}/share/${targetSchedule.share_id}`;
-    setShareUrl(url);
-    setShowShareDialog(true);
-  };
-
-  const handleImport = async (name: string, courses: Course[]) => {
-    await importSchedule(name, courses);
-    setShowImportDialog(false);
-  };
 
   return (
     <div className="max-w-7xl mx-auto py-8 px-4 sm:px-6 lg:px-8 min-h-screen">
