+++ lib/hooks/use-saved-schedules.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect, useCallback, useRef } from 'react';
 import {
   getSavedSchedules,
   deleteSavedSchedule,
@@ -9,14 +9,35 @@ import {
   type SavedSchedule,
 } from '@/lib/services/schedules';
 import { useMessage } from './use-message';
-import { Course } from '@/lib/types/course';
+import { useScheduleManagement } from './use-schedule-management';
+import type { Course } from '@/lib/types/course';
 
 export function useSavedSchedules() {
+  // --- Core State ---
   const [schedules, setSchedules] = useState<SavedSchedule[]>([]);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
-  const { showSuccess, showError } = useMessage();
 
+  // --- UI State ---
+  const [showDeleteAlert, setShowDeleteAlert] = useState(false);
+  const [scheduleToDelete, setScheduleToDelete] =
+    useState<SavedSchedule | null>(null);
+  const [activeSchedule, setActiveSchedule] = useState<SavedSchedule | null>(
+    null
+  );
+  const [showImportDialog, setShowImportDialog] = useState(false);
+  const [showShareDialog, setShowShareDialog] = useState(false);
+  const [shareUrl, setShareUrl] = useState('');
+
+  // --- Refs ---
+  const previewRef = useRef<HTMLDivElement>(null);
+
+  // --- Hooks ---
+  const { showSuccess, showError } = useMessage();
+  const { selectedCourses, conflicts, setSelectedCoursesDirectly } =
+    useScheduleManagement();
+
+  // --- Data Fetching ---
   const fetchSchedules = useCallback(async () => {
     try {
       setIsLoading(true);
@@ -34,11 +55,28 @@ export function useSavedSchedules() {
     fetchSchedules();
   }, [fetchSchedules]);
 
-  const handleDeleteSchedule = async (scheduleId: string) => {
+  // --- Effects ---
+  useEffect(() => {
+    if (activeSchedule) {
+      setSelectedCoursesDirectly(activeSchedule.schedule_data || []);
+      setTimeout(() => {
+        previewRef.current?.scrollIntoView({
+          behavior: 'smooth',
+          block: 'start',
+        });
+      }, 100);
+    } else {
+      setSelectedCoursesDirectly([]);
+    }
+  }, [activeSchedule, setSelectedCoursesDirectly]);
+
+  // --- Core Logic Handlers (internal) ---
+  const internalDeleteSchedule = async (scheduleId: string) => {
     try {
-      const scheduleName = schedules.find(s => s.id === scheduleId)?.schedule_name || 'Jadwal';
+      const scheduleName =
+        schedules.find((s) => s.id === scheduleId)?.schedule_name || 'Jadwal';
       await deleteSavedSchedule(scheduleId);
-      setSchedules(prev => prev.filter(s => s.id !== scheduleId));
+      setSchedules((prev) => prev.filter((s) => s.id !== scheduleId));
       showSuccess(`Jadwal "${scheduleName}" berhasil dihapus.`);
       return true;
     } catch (err) {
@@ -48,10 +86,10 @@ export function useSavedSchedules() {
     }
   };
 
-  const handleImportSchedule = async (name: string, courses: Course[]) => {
+  const internalImportSchedule = async (name: string, courses: Course[]) => {
     try {
       const newSchedule = await saveSchedule(name, courses);
-      setSchedules(prev => [newSchedule, ...prev]);
+      setSchedules((prev) => [newSchedule, ...prev]);
       showSuccess('Jadwal berhasil diimpor.');
       return newSchedule;
     } catch (err) {
@@ -61,11 +99,14 @@ export function useSavedSchedules() {
     }
   };
 
-  const handleUpdateSharing = async (scheduleId: string, isShared: boolean) => {
+  const internalUpdateSharing = async (
+    scheduleId: string,
+    isShared: boolean
+  ) => {
     try {
       const updatedSchedule = await updateScheduleSharing(scheduleId, isShared);
-      setSchedules(prev =>
-        prev.map(s => (s.id === scheduleId ? updatedSchedule : s)),
+      setSchedules((prev) =>
+        prev.map((s) => (s.id === scheduleId ? updatedSchedule : s))
       );
       if (isShared) {
         showSuccess('Jadwal sekarang publik dan bisa dibagikan.');
@@ -78,13 +119,92 @@ export function useSavedSchedules() {
     }
   };
 
+  // --- UI-Facing Event Handlers ---
+  const handleDeleteClick = (schedule: SavedSchedule) => {
+    setScheduleToDelete(schedule);
+    setShowDeleteAlert(true);
+  };
+
+  const handleConfirmDelete = async () => {
+    if (!scheduleToDelete) return;
+    const success = await internalDeleteSchedule(scheduleToDelete.id);
+    if (success && activeSchedule?.id === scheduleToDelete.id) {
+      setActiveSchedule(null);
+    }
+    setShowDeleteAlert(false);
+    setScheduleToDelete(null);
+  };
+
+  const handlePreviewClick = (schedule: SavedSchedule) => {
+    setActiveSchedule(schedule);
+  };
+
+  const closePreview = () => {
+    setActiveSchedule(null);
+  };
+
+  const handleExport = (schedule: SavedSchedule) => {
+    const exportData = {
+      type: 'planify-schedule',
+      version: 1,
+      scheduleName: schedule.schedule_name,
+      data: schedule.schedule_data,
+    };
+    const dataStr = JSON.stringify(exportData, null, 2);
+    const dataUri =
+      'data:application/json;charset=utf-8,' + encodeURIComponent(dataStr);
+    const exportFileDefaultName = `${schedule.schedule_name}.json`;
+    const linkElement = document.createElement('a');
+    linkElement.setAttribute('href', dataUri);
+    linkElement.setAttribute('download', exportFileDefaultName);
+    linkElement.click();
+  };
+
+  const handleShareClick = async (schedule: SavedSchedule) => {
+    let targetSchedule = schedule;
+    if (!targetSchedule.is_shared) {
+      const updatedSchedule = await internalUpdateSharing(
+        targetSchedule.id,
+        true
+      );
+      if (!updatedSchedule) return;
+      targetSchedule = updatedSchedule;
+    }
+    const url = `${window.location.origin}/share/${targetSchedule.share_id}`;
+    setShareUrl(url);
+    setShowShareDialog(true);
+  };
+
+  const handleImport = async (name: string, courses: Course[]) => {
+    await internalImportSchedule(name, courses);
+    setShowImportDialog(false);
+  };
+
   return {
+    // State and Refs
     schedules,
     isLoading,
     error,
-    deleteSchedule: handleDeleteSchedule,
-    importSchedule: handleImportSchedule,
-    updateSharing: handleUpdateSharing,
+    previewRef,
+    selectedCourses,
+    conflicts,
+    showDeleteAlert,
+    setShowDeleteAlert,
+    scheduleToDelete,
+    activeSchedule,
+    showImportDialog,
+    setShowImportDialog,
+    showShareDialog,
+    setShowShareDialog,
+    shareUrl,
+    // Action Handlers
+    handleDeleteClick,
+    handleConfirmDelete,
+    handlePreviewClick,
+    closePreview,
+    handleExport,
+    handleShareClick,
+    handleImport,
     refreshSchedules: fetchSchedules,
   };
-} 
+}
